Gameplay: enable minimap that zoom 2x out of the main camera

+++ engine/source/engine/window/Window.cpp
@@ -24,12 +24,12 @@ namespace gswy {
 	int Window::width = 1280;
 	int Window::height = 720;
 
-	int GSWY_GetWindowWidth()
+	float GSWY_GetWindowWidth()
 	{
 		return Window::width;
 	}
 
-	int GSWY_GetWindowHeight()
+	float GSWY_GetWindowHeight()
 	{
 		return Window::height;
 	}
